Moved base components to a dedicated folder Updated all imports of the base components
