+++ src/app/(pages)/(authenticated)/dashboard/loading.tsx
@@ -1,4 +1,4 @@
-import DashboardSkeleton from '@/ui/skeletons';
+import DashboardSkeleton from '@/ui/base-components/skeletons';
 
 export default function Loading() {
   return <DashboardSkeleton />;
