+++ src/ui/auth/signup-form.tsx
@@ -14,7 +14,7 @@ import {
   ExclamationCircleIcon,
 } from '@heroicons/react/24/outline';
 import { ArrowRightIcon } from '@heroicons/react/20/solid';
-import Button from '@/ui/button';
+import Button from '@/ui/base-components/button';
 import { useCallback, useState } from 'react';
 import { useFormStatus } from 'react-dom';
 import createUser from '@/lib/actions';
